+++ odoo/local-src/bso_hr_holiday_status/__manifest__.py
@@ -1,0 +1,13 @@
+{
+    'name': 'Hr holidays status',
+    'category': 'HR',
+    'description': 'Placeholder',
+    'author': 'BSO',
+    'website': 'https://www.bsonetwork.com',
+    'depends': ['base', 'hr_holidays'],
+    'data': [
+        'security/ir.model.access.csv',
+        'views/hr_holidays_status_view.xml',
+    ],
+    'application': False,
+}