+++ odoo/local-src/bso_hr_holiday_status/models/hr_holidays_status.py
@@ -1,0 +1,8 @@
+from odoo import models, fields
+
+
+class HrHolidaysStatus(models.Model):
+    _inherit = 'hr.holidays.status'
+
+    start_date = fields.Date()
+    end_date = fields.Date()